requirements.txt
django
djangorestframework
djanogrestframework_simplejwt
django-cors-headers
pillow